Bug Fixes
Im back.

Took a dive to why the program was bugging again (for f sake).

The reason for this was in Parser.py file - the articles generated didnt
have index = False which generated "Unamed" columns, which after some
time piled up and caused problems. This is why it was hard to find and
fix because i was looking at DataMatching.py for errors (A stupid
mistake, I admit).

I also made EVERY .to_csv have index = False, ran a couple of tests and
it SHOULD (I already hate this word) work now.

Also changed some of the syntax, i.e. endd -> end.